501 WGS - gl_auto data mendel errors
Mendel errors are corrected at the beginning of the process. This way, the MAF used in the clumping is more accurate. This causes the clumping procedure to selection new variants. Among them, a new problem occurs when running gl_auto: remaining mendel errors detected by the gl_auto software (which uses the whole pedigree to do so) that PLINK was not able to detect (using the .fam of the genotypes only) interrupt the procedure. A new section was added to gl_auto.sh to correct the situation.